+++ libs/@local/harpc/codec/Cargo.toml
@@ -11,6 +11,7 @@ publish.workspace = true
 [dependencies]
 # Public workspace dependencies
 error-stack = { workspace = true, public = true, features = ["serde"] }
+harpc-types = { workspace = true, public = true }
 
 # Public third-party dependencies
 bytes = { workspace = true, public = true }
@@ -18,14 +19,12 @@ futures-core = { workspace = true, public = true }
 serde = { workspace = true, public = true, features = ["derive"]}
 
 # Private workspace dependencies
-proptest = { workspace = true, optional = true }
-test-strategy = { workspace = true, optional = true }
+
+# Private third-party dependencies
 futures-util = { workspace = true, optional = true }
 serde_json = { workspace = true, optional = true, public = true }
 pin-project-lite = { workspace = true, optional = true }
 memchr = { workspace = true, optional = true }
-
-# Private third-party dependencies
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
@@ -34,5 +33,4 @@ tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 workspace = true
 
 [features]
-proptest = ["dep:proptest", "dep:test-strategy"]
 json = ["dep:serde_json", "dep:pin-project-lite", "dep:futures-util", "dep:memchr"]
