+++ libs/@local/harpc/net/Cargo.toml
@@ -10,6 +10,8 @@ authors.workspace = true
 
 [dependencies]
 # Public workspace dependencies
+harpc-codec = { workspace = true, public = true }
+
 
 # Public third-party dependencies
 futures-core = { workspace = true, public = true }
@@ -30,7 +32,7 @@ tokio = { workspace = true, public = true, features = ["io-util", "macros"] }
 codec = { workspace = true, features = ["harpc"] }
 error-stack = { workspace = true }
 harpc-wire-protocol = { workspace = true }
-harpc-codec = { workspace = true }
+harpc-types = { workspace = true }
 
 # Private third-party dependencies
 bytes = { workspace = true }
