+++ libs/@local/harpc/service/src/role.rs
@@ -1,7 +1,7 @@
 use core::marker::PhantomData;
 
 use harpc_tower::{body::Body, request::Request, response::Response};
-use harpc_wire_protocol::response::kind::ResponseKind;
+use harpc_types::response_kind::ResponseKind;
 
 mod sealed {
     pub trait Sealed {}
