+++ libs/@local/harpc/tower/src/net/unpack.rs
@@ -8,7 +8,7 @@ use core::{
 use bytes::Bytes;
 use futures::{Stream, StreamExt};
 use harpc_net::session::client::{ErrorStream, ResponseStream, TransactionStream, ValueStream};
-use harpc_wire_protocol::response::kind::ResponseKind;
+use harpc_types::response_kind::ResponseKind;
 
 use crate::body::{Body, BodyFrameResult, BodyState, Frame};
