+++ libs/@local/harpc/types/src/lib.rs
@@ -1,3 +1,7 @@
+#![feature(macro_metavar_expr, never_type)]
+
+pub mod error_code;
 pub mod procedure;
+pub mod response_kind;
 pub mod service;
 pub mod version;
