+++ libs/@local/harpc/types/src/procedure.rs
@@ -20,3 +20,10 @@ impl ProcedureId {
         self.0 & 0xF000 == 0xF000
     }
 }
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "proptest", derive(test_strategy::Arbitrary))]
+pub struct ProcedureDescriptor {
+    pub id: ProcedureId,
+}
