+++ libs/@local/harpc/types/src/service.rs
@@ -1,3 +1,5 @@
+use crate::version::Version;
+
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[cfg_attr(feature = "proptest", derive(test_strategy::Arbitrary))]
@@ -19,4 +21,12 @@ impl ServiceId {
         // 0xFxxx are reserved for internal use
         self.0 & 0xF000 == 0xF000
     }
+}
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "proptest", derive(test_strategy::Arbitrary))]
+pub struct ServiceDescriptor {
+    pub id: ServiceId,
+    pub version: Version,
 }
