+++ libs/@local/harpc/wire-protocol/Cargo.toml
@@ -27,7 +27,6 @@ thiserror = { workspace = true }
 proptest = { workspace = true }
 test-strategy = { workspace = true }
 harpc-types = { workspace = true, features = ["proptest"] }
-harpc-codec = { workspace = true, features = ["proptest"] }
 expect-test = { workspace = true }
 similar-asserts = { workspace = true }
