+++ libs/@local/harpc/wire-protocol/src/response/kind.rs
@@ -1,56 +1,8 @@
-use core::num::NonZero;
-
 use bytes::{Buf, BufMut};
 use error_stack::Result;
-use harpc_codec::error::ErrorCode;
+use harpc_types::response_kind::ResponseKind;
 
 use crate::codec::{Buffer, BufferError, Decode, Encode};
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
-#[cfg_attr(test, derive(test_strategy::Arbitrary))]
-pub enum ResponseKind {
-    Ok,
-    Err(ErrorCode),
-}
-
-impl ResponseKind {
-    #[must_use]
-    pub const fn is_ok(self) -> bool {
-        matches!(self, Self::Ok)
-    }
-
-    #[must_use]
-    pub const fn is_err(self) -> bool {
-        matches!(self, Self::Err(_))
-    }
-}
-
-impl From<u16> for ResponseKind {
-    fn from(value: u16) -> Self {
-        NonZero::new(value).map_or(Self::Ok, |value| Self::Err(ErrorCode::new(value)))
-    }
-}
-
-impl AsRef<Self> for ResponseKind {
-    fn as_ref(&self) -> &Self {
-        self
-    }
-}
-
-impl From<!> for ResponseKind {
-    fn from(never: !) -> Self {
-        never
-    }
-}
-
-impl From<ResponseKind> for u16 {
-    fn from(kind: ResponseKind) -> Self {
-        match kind {
-            ResponseKind::Ok => 0,
-            ResponseKind::Err(code) => code.value().get(),
-        }
-    }
-}
 
 impl Encode for ResponseKind {
     type Error = BufferError;
@@ -81,12 +33,9 @@ mod test {
     use core::num::NonZero;
 
     use expect_test::expect;
-    use harpc_codec::error::ErrorCode;
+    use harpc_types::{error_code::ErrorCode, response_kind::ResponseKind};
 
-    use crate::{
-        codec::test::{assert_codec, assert_decode, assert_encode},
-        response::kind::ResponseKind,
-    };
+    use crate::codec::test::{assert_codec, assert_decode, assert_encode};
 
     #[test]
     fn encode() {
